Working on making modules/core use a config dict instead of kwargs
This commit is the beginning of work to make the framework use a config
dict instead of kwargs to configure each module. This is because we can
then write any number of loader modules to load config anyway the end user
needs. I.e from file, from network, from a database.

It also cleans up somewhat the logging, to help isolate modules for
debugging.

This commit breaks all the modules as I haven't tested it yet. Checkout
v0.2 for a working aliasbot module that will show you latest features like
threading/indentity maps and so on